Moves the generator scripts to its own file
